requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
exceptiongroup==1.2.0
greenlet==3.0.1
idna==3.3
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pyee==11.0.1
pytest==7.4.4
pytest-base-url==2.0.0
python-slugify==8.0.1
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
text-unidecode==1.3
timeloop==1.0.2
tomli==2.0.1
tornado==6.1
typing_extensions==4.9.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
